Only await the status
I was running into timeouts while updating gadget to use the new npm
package. It turns out these `await call` lines were the issue.

The `await call` implementation does a `await Promise.all([this.headers,
this.status, this.trailers])`. After debugging, it seems
`this.headers` never resolves which was causing the timeouts.

We only care about the status anyway so I'm just gonna stick with this
for now.