Make the clients return a list of the files that changed after an update
Before this change, callers of the binary client know how many things changed, but not which things changed. There's a few bits of business logic gadget side that would be a lot easier if they knew exactly what changed, namely looking for changes to package.json during operation-runner commands, or looking for changes only to client-side files so we don't need to restart the server. This hands that information back to the client so they can work with it. We were already generating it so no huge performance hit there, but it will potentially be big when we send it back to the client so I think we may need to use this carefully.